% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_control_charts.R
\name{pca_control_charts}
\alias{pca_control_charts}
\title{Multivariate Control Charts based on Principal Components}
\usage{
pca_control_charts(x, alpha = 0.05, scaled = FALSE)
}
\arguments{
\item{x}{dataset.}

\item{alpha}{confidence level of the quality control charts.}

\item{scaled}{a logical value indicating whether the variables should be standardized.}
}
\description{
Draws a confidence ellipse based on the first two principal components and a multivariate quality control chart using the others principals components
}
\examples{
data(Overtime)
pca_control_charts(Overtime,alpha = .05,scaled = TRUE)

}
\author{
Gualberto Segundo Agamez Montalvo; Bruno Marinho Estevam de Oliveira
}
