% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Welder.R
\docType{data}
\name{Welder}
\alias{Welder}
\title{Control of Robotic Welders}
\format{
A data frame with 40 rows and 4 variables:
\describe{
  \item{X1}{Voltage (volts)}
  \item{X2}{Current (amps)}
  \item{X3}{Feed Speed (in/min)}
  \item{X4}{Gas Flow (cfm)}
}
}
\source{
Table 5.9 of Johnson and Wichern (2008) Applied Multivariate Statistical Analysis. 6th Edition, Prentice-Hall, New York.
}
\usage{
Welder
}
\description{
Entries for automated welding machines that must be controlled to be within
certain operational limits where a machine produces good quality welds.
}
\keyword{datasets}
